backend-ai/requirements.txt
fastapi
uvicorn
google-generativeai
psycopg2-binary
requests
yt-dlp